pyproject.toml
[project]
name = "ai-job-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apify-client>=2.1.0",
    "openai>=2.1.0",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
